+++ include/rtsplib/server/xop/MediaSession.h
@@ -61,7 +61,9 @@ public:
     { return _sessionId; }
 
     uint32_t getNumClient() const
-    { return _clients.size(); }
+    {
+		return static_cast<uint32_t>( _clients.size());
+	}
 
     bool isMulticast() const
     { return _isMulticast; }
